moving diff utils to utils/diffing
